requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
Flask-Script==2.0.6
google-cloud-secret-manager==2.0.0
python-dotenv==0.20.0
pandas==1.4.3
pandas-gbq==0.17.8
beautifulsoup4==4.11.1